+++ src/lib/components/utilities/browserrender/BrowserRender.svelte
@@ -1,0 +1,11 @@
+<script lang="ts">
+    import {IS_BROWSER} from "../../../util/environment";
+
+    type $$Slots = {
+        default: {};
+    };
+</script>
+
+{#if IS_BROWSER}
+    <slot />
+{/if}