+++ src/lib/util/environment.ts
@@ -2,3 +2,5 @@
 // `window` polyfilled, so this would throw a false positive
 
 export const IS_BROWSER: boolean = typeof window !== "undefined";
+
+export const IS_SERVER: boolean = !IS_BROWSER;